backend/requirements.txt
fastapi>=0.115
uvicorn>=0.34
pydantic>=2.0
google-cloud-storage>=2.18
